All migration in 1.0.0 handler
